capnproto: add lite_mode option
Cap'n Proto's lite mode disables all reflection and I/O features, so
that a lean library for encoding and decoding messages remains. This
especially is required for targeting Android with API level 21,
because libkj-async requires API level 23 or later for sigtimedwait.

It may be necessary to additionally add a tool_requires("capnproto")
when using lite mode, just like when cross-compiling, so that capnpc
is available to the lite build of libcapnp and libkj.